[step2] feat : 중복된 로또번호 검증 메소드, 예외처리 테스트 추가
